+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FIXES
 
+* Added a module id validator to fix issue [182](https://github.com/hashicorp/terraform-mcp-server/issues/182)
 * Fixes in readme for `TFE_HOSTNAME` v/s `TFE_ADDRESS`
 
 ## 0.3.1 (Oct 3, 2025)